refactor: #54 - Core모듈로 UIApplication Extension 이전, Alert창 출력 메서드 구현
