+++ Targets/Service/Sources/Extension/Reactive+.swift
@@ -7,7 +7,6 @@
 //
 
 import Core
-import DesignKit
 import Domain
 import UIKit
 
@@ -53,7 +52,7 @@ public extension Reactive where Base: MoyaProviderType {
             do {
               if !isValidationToken(try response.map(TokenErrorResponse.self)) {
                 TokenManager.deleteAll()
-                presentAlert()
+                UIApplication.presentTokenExpiredAlert()
                 return false
               }
             } catch {
@@ -104,23 +103,5 @@ public extension Reactive where Base: MoyaProviderType {
     else { return true }
 
     return false
-  }
-
-  private func presentAlert() {
-    let alertController: UIAlertController = .init(
-      title: "세션이 만료되었습니다",
-      message: "다시 로그인 후 시도해 주세요",
-      preferredStyle: .alert
-    )
-    alertController.modalPresentationStyle = .overCurrentContext
-    alertController.modalTransitionStyle = .crossDissolve
-    alertController.view.backgroundColor = .createColor(.monoGray, .w800)
-    alertController.view.cornerRadius(radius: 16)
-
-    alertController.addAction(.init(title: "확인", style: .default, handler: { _ in
-      NotificationCenter.default.post(name: .logout, object: nil)
-    }))
-
-    UIApplication.keyWindow?.rootViewController?.present(alertController, animated: true)
   }
 }
